virtio_drivers::device
Module socket

Driver for VirtIO socket devices.

To use the driver, you should first create a VirtIOSocket instance with your VirtIO transport, and then create a VsockConnectionManager wrapping it to keep track of connections. If you want to manage connections yourself you can use the VirtIOSocket directly for a lower-level interface.

See VsockConnectionManager for a usage example.

Structs

ConnectionInfo

Information about a particular vsock connection.

StreamShutdown

Flags sent with a shutdown request to hint that the peer won’t send or receive more data.

VirtIOSocket

Low-level driver for a VirtIO socket device.

VsockAddr

Socket address.

VsockConnectionManager

A higher level interface for VirtIO socket (vsock) devices.

VsockEvent

An event received from a VirtIO socket device.

Enums

DisconnectReason

The reason why a vsock connection was closed.

SocketError

The error type of VirtIO socket driver.

VsockEventType

Details of the type of an event received from a VirtIO socket.

Constants

VMADDR_CID_HOST

Well-known CID for the host.