virtio_drivers::device::socket

Enum SocketError

pub enum SocketError {

    ConnectionExists,
    ConnectionFailed,
    NotConnected,
    PeerSocketShutdown,
    NoResponseReceived,
    BufferTooShort,
    OutputBufferTooShort(usize),
    BufferTooLong(usize, usize),
    UnknownOperation(u16),
    InvalidOperation,
    InvalidNumber,
    UnexpectedDataInPacket,
    InsufficientBufferSpaceInPeer,
    RecycledWrongBuffer,

}

The error type of VirtIO socket driver.

Variants

ConnectionExists

There is an existing connection.

ConnectionFailed

Failed to establish the connection.

NotConnected

The device is not connected to any peer.

PeerSocketShutdown

Peer socket is shutdown.

NoResponseReceived

No response received.

BufferTooShort

The given buffer is shorter than expected.

OutputBufferTooShort(usize)

The given buffer for output is shorter than expected.

BufferTooLong(usize, usize)

The given buffer has exceeded the maximum buffer size.

UnknownOperation(u16)

Unknown operation.

InvalidOperation

Invalid operation,

InvalidNumber

Invalid number.

UnexpectedDataInPacket

Unexpected data in packet.

InsufficientBufferSpaceInPeer

Peer has insufficient buffer space, try again later.

RecycledWrongBuffer

Recycled a wrong buffer.