virtio_drivers::device::sound
Struct VirtIOSound

pub struct VirtIOSound<H: Hal, T: Transport> { /* private fields */ }

Audio driver based on virtio v1.2.

Supports synchronous blocking and asynchronous non-blocking audio playback.

Currently, only audio playback functionality has been implemented.

Implementations

impl<H: Hal, T: Transport> VirtIOSound<H, T>

pub fn new(transport: T) -> Result<Self>

Create a new VirtIO-Sound driver.

pub fn jacks(&self) -> u32

Total jack num.

pub fn streams(&self) -> u32

Total stream num.

pub fn chmaps(&self) -> u32

Total chmap num.

pub fn ack_interrupt(&mut self) -> bool

Acknowledge interrupt.

pub fn enable_interrupts(&mut self, enable: bool)

Enables interrupts from the device.

pub fn jack_remap( &mut self, jack_id: u32, association: u32, sequence: u32, ) -> Result

If the VIRTIO_SND_JACK_F_REMAP feature bit is set in the jack information, then the driver can send a control request to change the association and/or sequence number for the specified jack ID.

Arguments

jack_id - A u32 int which is in the range of [0, jacks)

pub fn pcm_set_params( &mut self, stream_id: u32, buffer_bytes: u32, period_bytes: u32, features: PcmFeatures, channels: u8, format: PcmFormat, rate: PcmRate, ) -> Result

Set selected stream parameters for the specified stream ID.

pub fn pcm_prepare(&mut self, stream_id: u32) -> Result

Prepare a stream with specified stream ID.

pub fn pcm_release(&mut self, stream_id: u32) -> Result

Release a stream with specified stream ID.

pub fn pcm_start(&mut self, stream_id: u32) -> Result

Start a stream with specified stream ID.

pub fn pcm_stop(&mut self, stream_id: u32) -> Result

Stop a stream with specified stream ID.

pub fn pcm_xfer(&mut self, stream_id: u32, frames: &[u8]) -> Result

Transfer PCM frame to device, based on the stream type(OUTPUT/INPUT).

Currently supports only output stream.

This is a blocking method that will not return until the audio playback is complete.

pub fn pcm_xfer_nb(&mut self, stream_id: u32, frames: &[u8]) -> Result<u16>

Transfer PCM frame to device, based on the stream type(OUTPUT/INPUT).

Currently supports only output stream.

This is a non-blocking method that returns a token.

The length of the frames must be equal to the buffer size set for the stream corresponding to the stream_id.

pub fn pcm_xfer_ok(&mut self, token: u16) -> Result

The PCM frame transmission corresponding to the given token has been completed.

pub fn output_streams(&mut self) -> Result<Vec<u32>>

Get all output streams.

pub fn input_streams(&mut self) -> Result<Vec<u32>>

Get all input streams.

pub fn rates_supported(&mut self, stream_id: u32) -> Result<PcmRates>

Get the rates that a stream supports.

pub fn formats_supported(&mut self, stream_id: u32) -> Result<PcmFormats>

Get the formats that a stream supports.

pub fn channel_range_supported( &mut self, stream_id: u32, ) -> Result<RangeInclusive<u8>>

Get channel range that a stream supports.

pub fn features_supported(&mut self, stream_id: u32) -> Result<PcmFeatures>

Get features that a stream supports.

pub fn latest_notification(&mut self) -> Result<Option<Notification>>

Get the latest notification.