Struct virtio_drivers::VirtIOInput

pub struct VirtIOInput<'a, H: Hal> { /* private fields */ }

Virtual human interface devices such as keyboards, mice and tablets.

An instance of the virtio device represents one such input device. Device behavior mirrors that of the evdev layer in Linux, making pass-through implementations on top of evdev easy.

Implementations

impl<'a, H: Hal> VirtIOInput<'a, H>

pub fn new(header: &'static mut VirtIOHeader) -> Result<Self>

Create a new VirtIO-Input driver.

pub fn ack_interrupt(&mut self) -> bool

Acknowledge interrupt and process events.

pub fn pop_pending_event(&mut self) -> Option<InputEvent>

Pop the pending event.

pub fn query_config_select(
&mut self,
select: InputConfigSelect,
subsel: u8,
out: &mut [u8]
) -> u8

Query a specific piece of information by select and subsel, and write result to out, return the result size.