virtio_drivers::transport::mmio

Struct MmioTransport

pub struct MmioTransport { /* private fields */ }

MMIO Device Register Interface.

Ref: 4.2.2 MMIO Device Register Layout and 4.2.4 Legacy interface

Implementations

impl MmioTransport

pub unsafe fn new(header: NonNull<VirtIOHeader>) -> Result<Self, MmioError>

Constructs a new VirtIO MMIO transport, or returns an error if the header reports an unsupported version.

Safety

header must point to a properly aligned valid VirtIO MMIO region, which must remain valid for the lifetime of the transport that is returned.

pub fn version(&self) -> MmioVersion

Gets the version of the VirtIO MMIO transport.

pub fn vendor_id(&self) -> u32

Gets the vendor ID.