virtio_drivers::transport::pci::bus

Enum BarInfo

pub enum BarInfo {
    Memory {
        address_type: MemoryBarType,
        prefetchable: bool,
        address: u64,
        size: u32,
    },
    IO {
        address: u32,
        size: u32,
    },
}

Information about a PCI Base Address Register.

Variants

Memory

The BAR is for a memory region.

Fields

address_type: MemoryBarType

The size of the BAR address and where it can be located.

prefetchable: bool

If true, then reading from the region doesnโt have side effects. The CPU may cache reads and merge repeated stores.

address: u64

The memory address, always 16-byte aligned.

size: u32

The size of the BAR in bytes.

IO

The BAR is for an I/O region.

Fields

address: u32

The I/O address, always 4-byte aligned.

size: u32

The size of the BAR in bytes.

Implementations

impl BarInfo

pub fn takes_two_entries(&self) -> bool

Returns whether this BAR is a 64-bit memory region, and so takes two entries in the table in configuration space.

pub fn memory_address_size(&self) -> Option<(u64, u32)>

Returns the address and size of this BAR if it is a memory bar, or None if it is an IO BAR.