virtio_drivers::transport::pci

Enum VirtioPciError

pub enum VirtioPciError {
    InvalidVendorId(u16),
    MissingCommonConfig,
    MissingNotifyConfig,
    InvalidNotifyOffMultiplier(u32),
    MissingIsrConfig,
    UnexpectedIoBar,
    BarNotAllocated(u8),
    BarOffsetOutOfRange,
    Misaligned {
        vaddr: NonNull<u8>,
        alignment: usize,
    },
    Pci(PciError),
}

An error encountered initialising a VirtIO PCI transport.

Variants

InvalidVendorId(u16)

PCI device vender ID was not the VirtIO vendor ID.

MissingCommonConfig

No valid VIRTIO_PCI_CAP_COMMON_CFG capability was found.

MissingNotifyConfig

No valid VIRTIO_PCI_CAP_NOTIFY_CFG capability was found.

InvalidNotifyOffMultiplier(u32)

VIRTIO_PCI_CAP_NOTIFY_CFG capability has a notify_off_multiplier that is not a multiple of 2.

MissingIsrConfig

No valid VIRTIO_PCI_CAP_ISR_CFG capability was found.

UnexpectedIoBar

An IO BAR was provided rather than a memory BAR.

BarNotAllocated(u8)

A BAR which we need was not allocated an address.

BarOffsetOutOfRange

The offset for some capability was greater than the length of the BAR.

Misaligned

The virtual address was not aligned as expected.

Fields

vaddr: NonNull<u8>

The virtual address in question.

alignment: usize

The expected alignment in bytes.

Pci(PciError)

A generic PCI error,