Crate virtio_drivers

VirtIO guest drivers.

Structs

BlkResp

Response of a VirtIOBlk request.

InputEvent

Both queues use the same virtio_input_event struct. type, code and value are filled according to the Linux input layer (evdev) interface.

VirtIOBlk

The virtio block device is a simple virtual block device (ie. disk).

VirtIOConsole

Virtio console. Only one single port is allowed since ``alloc’’ is disabled. Emergency and cols/rows unimplemented.

VirtIOGpu

A virtio based graphics adapter.

VirtIOHeader

MMIO Device Legacy Register Interface.

VirtIOInput

Virtual human interface devices such as keyboards, mice and tablets.

VirtIONet

The virtio network device is a virtual ethernet card.

Enums

DeviceType

Types of virtio devices.

Error

The error type of VirtIO drivers.

InputConfigSelect

Select value used for VirtIOInput::query_config_select().

RespStatus

Status of a VirtIOBlk request.

Traits

Hal

The interface which a particular hardware implementation must implement.

Type Definitions

PhysAddr

A physical address as used for virtio.

Result

The type returned by driver methods.

VirtAddr

A virtual memory address in the address space of the program.