Trait virtio_drivers::Hal

pub trait Hal {
    fn dma_alloc(pages: usize) -> PhysAddr;
    fn dma_dealloc(paddr: PhysAddr, pages: usize) -> i32;
    fn phys_to_virt(paddr: PhysAddr) -> VirtAddr;
    fn virt_to_phys(vaddr: VirtAddr) -> PhysAddr;
}

The interface which a particular hardware implementation must implement.

Required Methods

fn dma_alloc(pages: usize) -> PhysAddr

Allocates the given number of contiguous physical pages of DMA memory for virtio use.

fn dma_dealloc(paddr: PhysAddr, pages: usize) -> i32

Deallocates the given contiguous physical DMA memory pages.

fn phys_to_virt(paddr: PhysAddr) -> VirtAddr

Converts a physical address used for virtio to a virtual address which the program can access.

fn virt_to_phys(vaddr: VirtAddr) -> PhysAddr

Converts a virtual address which the program can access to the corresponding physical address to use for virtio.